third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/ShadowRoot-init-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] attachShadow() should use null registry when customElementRegistry is null (host uses global registry)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] attachShadow() should use null registry when customElementRegistry is null (host uses custom registry)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
